# docs/api/newsletter-openapi.yaml
openapi: 3.0.3
info:
  title: Newsletter API
  description: |
    Complete newsletter subscription management API for Next.js applications.
    
    Features:
    - Email subscription and unsubscription
    - Double opt-in confirmation
    - Analytics and statistics
    - Bulk operations
    - Multiple database backend support
    - Rate limiting and security
    
    Built for modern web applications with TypeScript support.
  version: 2.0.0
  contact:
    name: API Support
    email: api-support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.example.com/newsletter
    description: Production server
  - url: https://api-staging.example.com/newsletter
    description: Staging server
  - url: http://localhost:3000/api/newsletter
    description: Development server
security:
  - ApiKeyAuth: []
  - BearerAuth: []
  - {}
paths:
...
  /unsubscribe:
    post:
      tags:
        - Subscriptions
      summary: Unsubscribe from newsletter
      description: |
        Remove an email address from the newsletter. Can use either email address
        or unsubscribe token. Token method is preferred for security.
      operationId: unsubscribe
      security:
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnsubscribeRequest'
            examples:
              with_token:
                summary: Using unsubscribe token
                value:
                  token: "unsubscribe-token-123"
                  reason: "too_frequent"
                  feedback: "Getting too many emails"
              with_email:
                summary: Using email address
                value:
                  email: "user@example.com"
                  reason: "not_relevant"
      responses:
        '200':
          description: Successfully unsubscribed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnsubscribeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Email not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                not_found:
                  summary: Email not found
                  value:
                    error: "Email not found in subscription list"
                    code: "EMAIL_NOT_FOUND"
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalError'
  /confirm:
    post:
      tags:
        - Subscriptions
      summary: Confirm subscription
      description: |
        Confirm a pending email subscription using the confirmation token
        sent via email (double opt-in).
      operationId: confirm
      security:
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmRequest'
      responses:
        '200':
          description: Subscription confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for admin authentication
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for admin authentication
  headers:
    X-RateLimit-Limit:
      description: The number of allowed requests in the current period
      schema:
        type: integer
        example: 100
...
    X-RateLimit-Reset:
      description: The time at which the current rate limit window resets (Unix timestamp)
      schema:
        type: integer
        example: 1640995200
  schemas:
...
    UnsubscribeRequest:
      type: object
      oneOf:
        - required: [token]
        - required: [email]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          description: Email address to unsubscribe
          example: "user@example.com"
        token:
          type: string
          description: Unsubscribe token from email
          example: "unsubscribe-token-123"
        reason:
          type: string
          enum: [too_frequent, not_relevant, never_signed_up, temporary, other]
          description: Reason for unsubscribing
          example: "too_frequent"
        feedback:
          type: string
          maxLength: 500
          description: Optional feedback text
          example: "Getting too many emails"
    UnsubscribeResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Successfully unsubscribed"
        data:
          type: object
          properties:
            email:
              type: string
              example: "user@example.com"
            unsubscribed_at:
              type: string
              format: date-time
              example: "2024-01-01T00:00:00.000Z"
            reason:
              type: string
              example: "too_frequent"
            can_resubscribe:
              type: boolean
              example: true
    ConfirmRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          description: Confirmation token from email
          example: "confirmation-token-123"
    ConfirmResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Subscription confirmed successfully"
        data:
          type: object
          properties:
            email:
              type: string
              example: "user@example.com"
            confirmed_at:
              type: string
              format: date-time
              example: "2024-01-01T00:00:00.000Z"
            status:
              type: string
              example: "active"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
          example: "Validation failed"
        code:
          type: string
          description: Machine-readable error code
          example: "VALIDATION_ERROR"
        details:
          type: object
          description: Additional error details
          example:
            field: "email"
            message: "Please enter a valid email address"
        timestamp:
          type: string
          format: date-time
          description: When the error occurred
          example: "2024-01-01T00:00:00.000Z"
        path:
          type: string
          description: Request path that caused the error
          example: "/api/newsletter/subscribe"
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation_error:
              summary: Validation error
              value:
                error: "Validation failed"
                code: "VALIDATION_ERROR"
                details:
                  email: "Please enter a valid email address"
                timestamp: "2024-01-01T00:00:00.000Z"
                path: "/api/newsletter/subscribe"
...
    RateLimitExceeded:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          schema:
            type: integer
            example: 0
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'
        Retry-After:
          description: Number of seconds to wait before retrying
          schema:
            type: integer
            example: 300
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            rate_limit:
              summary: Rate limit exceeded
              value:
                error: "Rate limit exceeded. Please wait before trying again."
                code: "RATE_LIMIT_EXCEEDED"
                details:
                  reset_time: 1640995200
                  retry_after: 300
                timestamp: "2024-01-01T00:00:00.000Z"
                path: "/api/newsletter/subscribe"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            internal_error:
              summary: Internal server error
              value:
                error: "Internal server error"
                code: "INTERNAL_ERROR"
                timestamp: "2024-01-01T00:00:00.000Z"
                path: "/api/newsletter/subscribe"
...